requirements.txt
atlassian-python-api
html2text
beautifulsoup4
